+++ src/method_versions/method.v0.5.ts
@@ -14,8 +14,9 @@ import type {
   UpdateDIDInterface,
   WitnessProofFileEntry,
 } from '../interfaces';
-import { createDate, deepClone, normalizeDidAddress, replaceValueInObject } from '../utils';
+import { deepClone, normalizeDidAddress, replaceValueInObject } from '../utils';
 import { createSCID, deriveHash } from '../utils/crypto';
+import { createDate } from '../utils/iso8601-datetime';
 import { countVerifiedWitnessApprovals, fetchWitnessProofs, validateWitnessParameter } from '../witness';
 
 const VERSION = '0.5';
