+++ src/utils.ts
@@ -653,12 +653,3 @@ export function replaceValueInObject<T>(obj: T, searchValue: string, replaceValu
   }
   return obj;
 }
-
-export const createDate = (created?: Date | string) =>
-  new Date(created ?? Date.now()).toISOString().replace(/\.\d{1,3}Z$/, 'Z');
-
-export function bytesToHex(bytes: Uint8Array): string {
-  return Array.from(bytes)
-    .map((byte) => byte.toString(16).padStart(2, '0'))
-    .join('');
-}
